requirements.txt
google-genai
google-cloud-storage
google-cloud-aiplatform
google-cloud-texttospeech
Pillow
opencv-python
requests
ffmpeg-python
controlnet_aux
torch
torchvision
torchaudio
mediapipe